fix: 解决循环导入问题并清理工具参数描述
1. 抽取 get_local_tools_description_zh 到独立模块 tool_local_service.py，避免循环导入
2. ToolInfo 模型添加 params: List 字段
3. 移除用户不感知参数(exclude=True)的 description_zh
4. 修复 send_email_tool description 中的 typo
