+++ backend/database/tool_db.py
@@ -4,7 +4,7 @@ from typing import List
 from database.agent_db import logger
 from database.client import get_db_session, filter_property, as_dict
 from database.db_models import ToolInstance, ToolInfo
-from consts.model import ToolSourceEnum
+from services.tool_local_service import get_local_tools_description_zh
 
 
 def create_tool(tool_info, version_no: int = 0):
@@ -37,7 +37,7 @@ def create_or_update_tool_by_tool_info(tool_info, tenant_id: str, user_id: str, 
     Args:
         tool_info: Dictionary containing tool information
         tenant_id: Tenant ID for filtering, mandatory
-        user_id: User ID for updating (will be set as the last updater)
+        user_id: Optional user ID for filtering
         version_no: Version number to filter. Default 0 = draft/editing state
 
     Returns:
@@ -48,10 +48,9 @@ def create_or_update_tool_by_tool_info(tool_info, tenant_id: str, user_id: str, 
 
     with get_db_session() as session:
         # Query if there is an existing ToolInstance
-        # Note: Do not filter by user_id to avoid creating duplicate instances
-        # for the same agent_id and tool_id when different users save
         query = session.query(ToolInstance).filter(
             ToolInstance.tenant_id == tenant_id,
+            ToolInstance.user_id == user_id,
             ToolInstance.agent_id == tool_info_dict['agent_id'],
             ToolInstance.delete_flag != 'Y',
             ToolInstance.tool_id == tool_info_dict['tool_id'],
@@ -64,12 +63,7 @@ def create_or_update_tool_by_tool_info(tool_info, tenant_id: str, user_id: str, 
                 if hasattr(tool_instance, key):
                     setattr(tool_instance, key, value)
         else:
-            # Create a new ToolInstance
-            new_tool_instance = ToolInstance(
-                **filter_property(tool_info_dict, ToolInstance))
-            session.add(new_tool_instance)
-            session.flush()  # Flush to get the ID
-            tool_instance = new_tool_instance
+            create_tool(tool_info_dict, version_no)
         return tool_instance
 
 
@@ -197,23 +191,13 @@ def check_tool_list_initialized(tenant_id: str) -> bool:
 def update_tool_table_from_scan_tool_list(tenant_id: str, user_id: str, tool_list: List[ToolInfo]):
     """
     scan all tools and update the tool table in PG database, remove the duplicate tools
-    For MCP tools, use name&source&usage as unique key to allow same tool name from different MCP servers
     """
     with get_db_session() as session:
         # get all existing tools (including complete information)
         existing_tools = session.query(ToolInfo).filter(ToolInfo.delete_flag != 'Y',
                                                         ToolInfo.author == tenant_id).all()
-        # Build existing_tool_dict with different keys for MCP vs non-MCP tools
-        existing_tool_dict = {}
-        for tool in existing_tools:
-            if tool.source == ToolSourceEnum.MCP.value:
-                # For MCP tools, use name + source + usage (MCP server name) as unique key
-                key = f"{tool.name}&{tool.source}&{tool.usage or ''}"
-            else:
-                # For other tools, use name + source as unique key
-                key = f"{tool.name}&{tool.source}"
-            existing_tool_dict[key] = tool
-
+        existing_tool_dict = {
+            f"{tool.name}&{tool.source}": tool for tool in existing_tools}
         # set all tools to unavailable
         for tool in existing_tools:
             tool.is_available = False
@@ -225,15 +209,9 @@ def update_tool_table_from_scan_tool_list(tenant_id: str, user_id: str, tool_lis
             is_available = True if re.match(
                 r'^[a-zA-Z_][a-zA-Z0-9_]*$', tool.name) is not None else False
 
-            # Use same key generation logic as above
-            if tool.source == ToolSourceEnum.MCP.value:
-                tool_key = f"{tool.name}&{tool.source}&{tool.usage or ''}"
-            else:
-                tool_key = f"{tool.name}&{tool.source}"
-
-            if tool_key in existing_tool_dict:
-                # by tool name, source, and usage (for MCP) to update the existing tool
-                existing_tool = existing_tool_dict[tool_key]
+            if f"{tool.name}&{tool.source}" in existing_tool_dict:
+                # by tool name and source to update the existing tool
+                existing_tool = existing_tool_dict[f"{tool.name}&{tool.source}"]
                 for key, value in filtered_tool_data.items():
                     setattr(existing_tool, key, value)
                 existing_tool.updated_by = user_id
@@ -248,8 +226,6 @@ def update_tool_table_from_scan_tool_list(tenant_id: str, user_id: str, tool_lis
 
 
 def add_tool_field(tool_info):
-    from services.tool_configuration_service import get_local_tools_description_zh
-    
     with get_db_session() as session:
         # Query if there is an existing ToolInstance
         query = session.query(ToolInfo).filter(
@@ -360,7 +336,6 @@ def delete_tools_by_agent_id(agent_id, tenant_id, user_id, version_no: int = 0):
         ).update({
             ToolInstance.delete_flag: 'Y', 'updated_by': user_id
         })
-
 
 def search_last_tool_instance_by_tool_id(tool_id: int, tenant_id: str, user_id: str, version_no: int = 0):
     """
